Add files via upload
When we run the code in Visual Studio a new window will appear (this window is created using Tkinter). We have to click on the start button then questions will appear which are stored in questions.txt file attached above. We have to choose the option that we think is correct. Lastly, it will calculate our score and display it.